refactor(components): refactor components
implement #46
