+++ src/components/NavbarPlaceholder.vue
