+++ src/components/icons/IconSuccess.vue
