+++ src/components/icons/IconWarning.vue
